+++ core/app/models/spree/legacy_user.rb
@@ -1,8 +1,8 @@
 # Default implementation of User.  This class is intended to be modified by extensions (ex. spree_auth_devise)
 module Spree
   class LegacyUser < Spree::Base
-    include Core::UserAddress
-    include Core::UserPaymentSource
+    include UserAddress
+    include UserPaymentSource
 
     self.table_name = 'spree_users'